pyproject.toml
[tool.poetry]
name = "tickzero"
version = "0.1.0"
description = "CS2 Highlight Generator"
authors = ["TickZero Team <team@tickzero.com>"]
readme = "README.md"
packages = [{include = "tickzero", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
pydantic = "^2.0"
google-genai = "^0.2.0"
obs-websocket-py = "^1.0"
ffmpeg-python = "*"
google-generativeai = "^0.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
nox = "*"
ruff = "*"
mypy = "*"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
strict = true
python_version = "3.10"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
